Refactor code editor utils into react hook
